Add files via upload
Using this code we can get the records mentioned in the csv file and then writing multiple SQL queries, join them basis the common field. Kindly note we are doing this activity because MSSQL server won't handle the load to create a new table. Hence we are taking the records for TID from the csv file and joining the sql queries from other tables to create a new table in sql for further analysis and BI dashboards.